implimenting the blob seems to be more difficult than first realised.
GA mutation Operator should randomly select components and shuffle them based on mutation Rate.

Blob should look at dependencies and associations, adding the contracts between components to the "MutationHeuristic" class, which is then passed to the overloaded GA Mutation Operator. Some way to increase
the mutation rate based on contracts should be used by the GA to increase the likelihood of mutating contracted
components together. RelationshipMatrix (renamed so not to be confused with assocations = composite and aggregate) may require changing to differentiate between different types of relationships. This class can then be used in place of the algorithms currently in blob - saving the blob having to do the processing of components and separating concerns more neatly.