work toward selection is mostly done but not quite finished, it requires cohesion fitness metrics to be completed, as it does not work as intended yet, as well as a way to work with the weighted methods per class once it is passed to the selection algorithm. So far the idea has been to pull the best of each of the metrics into a seporate pool and then randomly select from these pools for offspring.
This is proving challenging though as knowing the cohesion relies on assumptions about the Collections created initially as being the attributes accessed by operations. Which will cause problems down the line without sequence diagrams to clarify behaviour or human interaction to determine at runtime what the software is supposed to do at each step which seems unlikely.

the program also seems to freeze or crash after pressing the evolution button in the GUI at runtime. This will need to be investigated.