added some primitive data type operation dependency exceptions. Now the program doesnt search for the majority of core or primitive data types when looking at operation parameter dependencies which is good.
Seemed to have fixed the cohesion fitness measure but will need looking over in finer detail with testing.

distribution of methods is a little brittle when tested with one of the old test cases. Not too bothered by this for the moment but it should be examined in more detail once the rest of the code is in a good condition.

names still arent appearing on the fitness bar graph upon the initial loading of the file. It's a little odd but I think it can be fixed easily enough with a line or two.